Add GRPO+SFT training entrypoint and supporting changes
- train_grpo_sft.py: GRPO+SFT joint training with DTensor/TP=2 support,
  bucketed sequence lengths, and configurable alpha_grpo/alpha_sft weights
- load_sft_jsonl: new dataset loader for SFT JSONL files
- GRPO+SFT configs for IFEval (baseline and v2 hyperparams)
- Deepscaler config, eval_pass_at_k --max-prompts flag, MATH ablation notes
- Fix wire_dtype float16->bfloat16 in train_grpo.py, fix data.py S3 cache race

+++ entrypoints/train_grpo.py
@@ -686,7 +686,7 @@ async def main():
                     trainer, reference,
                     abort_in_flight=abort_in_flight,
                     trainer_version=trainer.version,
-                    wire_dtype="float16",
+                    wire_dtype="bfloat16",
                 )
 
             # Sync weights to rollout model
@@ -696,7 +696,7 @@ async def main():
                     trainer, rollout,
                     abort_in_flight=abort_in_flight,
                     trainer_version=trainer.version,
-                    wire_dtype="float16",
+                    wire_dtype="bfloat16",
                 )
 
             # Checkpoint (saves to SM_MODEL_DIR on SageMaker, uploaded at job end)
